Change the balance column to integer to avoid floating point errors

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, String
+from sqlalchemy import Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 
 from .database import Base
@@ -8,7 +8,7 @@ class Wallet(Base):
     __tablename__ = "wallets"
 
     uuid = Column(String, primary_key=True, index=True, unique=True)
-    balance = Column(Float, default=0.0)
+    balance = Column(Integer, default=0)
 
 
 class WalletError(Exception):
